\name{distillery internal}
\alias{print.booted}
\alias{summary.booted}
\alias{print.ci.booted}
\alias{print.tibbed}
\alias{print.tibRMed}
\alias{plot.tibbed}
\alias{plot.tibRMed}
\title{
  distillery Internal and Secondary Functions
}

\description{
Listed below are supporting functions for the major methods in distillery.
}

\keyword{internal}
\keyword{misc}
