% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distillery-package.R
\docType{package}
\name{distillery-package}
\alias{distillery}
\alias{distillery-package}
\title{distillery: Compress arbitrary mlr3 models into keras models}
\description{
Compresses arbitrary keras models into neural networks.
    Methodology similar to Caruana, 2015: Model Compression.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3misc.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3misc}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3misc/issues}
}

}
\author{
\strong{Maintainer}: Jakob Bodensteiner \email{jakob.bodensteiner@campus.lmu.de}

Authors:
\itemize{
  \item Florian Pfisterer \email{florian.pfisterer@stat.uni-muenchen.de}
}

}
