% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{gen_artificial_data}
\alias{gen_artificial_data}
\title{Generate artificial data, keeping 'keep_fraction' bootstrap samples.
Keeps the original label}
\usage{
gen_artificial_data(task, min_n = 10000, keep_fraction = 0.1, ...)
}
\arguments{
\item{task}{[`Task ]
Task to create artificial samples from.}

\item{min_n}{[`integer`] \cr
Minimum number of observations to create. Default: 10000.}

\item{keep_fraction}{[`numeric`] \cr
Percentage of original data fraction to keep. Default: 0.1.}

\item{...}{`any` \cr
Passed on to `munge_task`.}
}
\description{
Generate artificial data, keeping 'keep_fraction' bootstrap samples.
Keeps the original label
}
