% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{make_munge_generators}
\alias{make_munge_generators}
\title{Create train / validation data generators from a task and params}
\usage{
make_munge_generators(
  task,
  x_transform,
  y_transform,
  validation_split = 1/3,
  batch_size = 128L,
  seed = NULL,
  ...
)
}
\arguments{
\item{task}{[`Task`]\cr
An mlr3 [`Task`].}

\item{x_transform}{[`function`]\cr
Function used to transform data to a keras input format for features.}

\item{y_transform}{[`function`]\cr
Function used to transform data to a keras input format for the response.}

\item{validation_split}{[`numeric`]\cr
Fraction of data to use for validation.}

\item{batch_size}{[`integer`]\cr
Batch_size for the generators.}

\item{...}{`any`\cr
Additional arguments passed on to `gen_artificial_data`}
}
\description{
Create train / validation data generators from a task and params
}
