% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{munge_task}
\alias{munge_task}
\title{Munge a Task.}
\usage{
munge_task(task, switch_prob = 0.5, var = 2)
}
\arguments{
\item{task}{[`Task ]
Task to create artificial samples from.}

\item{switch_prob}{[`numeric`] \cr
Swapping probabilitiy. Default: 0.5.}

\item{var}{[`numeric`] \cr
For numeric features, a new value is sampled from N(x_nn, |x -x_nn| / var)
where x is a feature value, x_nn the nearest neighbour's corresponding feature value.
Default: 2.}
}
\description{
Munge a Task.
}
\details{
Always creates a new dataset exactly as large as 'task'.
}
